# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareConnect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  icon: string;
  caregiverCount: number;
}

export interface Caregiver {
  id: number;
  name: string;
  bio: string;
  avatarUrl?: string | null;
  location: string;
  categoryIds: number[];
  categories?: Category[];
  hourlyRate: number;
  rating: number;
  reviewCount: number;
  yearsExperience: number;
  isVerified: boolean;
  createdAt: string;
}

export interface CreateCaregiverBody {
  name: string;
  bio: string;
  location: string;
  categoryIds: number[];
  hourlyRate: number;
  yearsExperience: number;
}

export type CareRequestStatus =
  (typeof CareRequestStatus)[keyof typeof CareRequestStatus];

export const CareRequestStatus = {
  open: "open",
  filled: "filled",
  cancelled: "cancelled",
} as const;

export interface CareRequest {
  id: number;
  title: string;
  description: string;
  categoryId: number;
  category?: Category;
  location: string;
  startDate: string;
  endDate?: string | null;
  budget: number;
  status: CareRequestStatus;
  seekerName: string;
  createdAt: string;
}

export interface CreateCareRequestBody {
  title: string;
  description: string;
  categoryId: number;
  location: string;
  startDate: string;
  endDate?: string;
  budget: number;
  seekerName: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  caregiverId: number;
  caregiver?: Caregiver;
  careRequestId: number;
  careRequest?: CareRequest;
  status: BookingStatus;
  message?: string;
  createdAt: string;
}

export interface CreateBookingBody {
  caregiverId: number;
  careRequestId: number;
  message?: string;
}

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
}

export type StatsOverviewCategoryCountsItem = {
  categoryName: string;
  count: number;
};

export interface StatsOverview {
  totalCaregivers: number;
  totalCareRequests: number;
  totalBookings: number;
  openRequests: number;
  categoryCounts: StatsOverviewCategoryCountsItem[];
}

export type ListCaregiversParams = {
  category?: string;
  location?: string;
  minRating?: number;
};

export type ListCareRequestsParams = {
  category?: string;
  status?: string;
};

export type ListBookingsParams = {
  caregiverId?: number;
  status?: string;
};