# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Caregiver } from "./caregiver";
import type { CareRequest } from "./careRequest";

export interface Booking {
  id: number;
  caregiverId: number;
  caregiver?: Caregiver;
  careRequestId: number;
  careRequest?: CareRequest;
  status: BookingStatus;
  message?: string;
  createdAt: Date;
}