# lib/api-zod/src/generated/types/caregiver.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface Caregiver {
  id: number;
  name: string;
  bio: string;
  avatarUrl?: string | null;
  location: string;
  categoryIds: number[];
  categories?: Category[];
  hourlyRate: number;
  rating: number;
  reviewCount: number;
  yearsExperience: number;
  isVerified: boolean;
  createdAt: Date;
}