# lib/api-zod/src/generated/types/createCareRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCareRequestBody {
  title: string;
  description: string;
  categoryId: number;
  location: string;
  startDate: string;
  endDate?: string;
  budget: number;
  seekerName: string;
}