# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatsOverviewCategoryCountsItem } from "./statsOverviewCategoryCountsItem";

export interface StatsOverview {
  totalCaregivers: number;
  totalCareRequests: number;
  totalBookings: number;
  openRequests: number;
  categoryCounts: StatsOverviewCategoryCountsItem[];
}