# lib/api-zod/src/generated/types/updateBookingStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CareConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBookingStatusBodyStatus } from "./updateBookingStatusBodyStatus";

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
}